dbeaver/pro#8907: renamed ce winstore bundle
