+++ features/org.jkiss.dbeaver.appstore.feature/feature.xml
@@ -20,6 +20,6 @@
     <plugin id="org.jkiss.dbeaver.ui.config.sample" version="0.0.0" />
 
     <includes id="org.jkiss.dbeaver.ui.extra.feature" version="0.0.0"/>
-    <plugin id="org.jkiss.dbeaver.ui.app.winstore" os="win32" arch="x86_64" version="0.0.0" />
+    <plugin id="org.jkiss.dbeaver.ui.app.msstore" os="win32" arch="x86_64" version="0.0.0"/>
 
 </feature>
