+++ plugins/org.jkiss.dbeaver.ui.app.msstore/META-INF/MANIFEST.MF
@@ -2,12 +2,12 @@ Manifest-Version: 1.0
 Bundle-ManifestVersion: 2
 Bundle-Vendor: DBeaver Corp
 Bundle-Name: DBeaver Windows Store extension
-Bundle-SymbolicName: org.jkiss.dbeaver.ui.app.winstore;singleton:=true
+Bundle-SymbolicName: org.jkiss.dbeaver.ui.app.msstore;singleton:=true
 Bundle-ActivationPolicy: lazy
 Bundle-Version: 1.0.102.qualifier
 Bundle-Release-Date: 20260420
 Bundle-RequiredExecutionEnvironment: JavaSE-21
-Export-Package: org.jkiss.dbeaver.winstore
+Export-Package: org.jkiss.dbeaver.msstore
 Require-Bundle: org.jkiss.dbeaver.model
-Automatic-Module-Name: org.jkiss.dbeaver.ui.app.winstore
+Automatic-Module-Name: org.jkiss.dbeaver.ui.app.msstore
