+++ plugins/org.jkiss.dbeaver.ui.app.msstore/build.properties
